Add aws-smithy-dns crate
