+++ rust-runtime/Cargo.toml
@@ -7,6 +7,7 @@ members = [
     "aws-smithy-checksums",
     "aws-smithy-compression",
     "aws-smithy-client",
+    "aws-smithy-dns",
     "aws-smithy-eventstream",
     "aws-smithy-http",
     "aws-smithy-http-auth",
