+++ rust-runtime/aws-smithy-dns/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "aws-smithy-dns"
+version = "0.1.0"
+authors = [
+    "AWS Rust SDK Team <aws-sdk-rust@amazon.com>",
+]
+description = "DNS resolvers for smithy-rs"
+edition = "2021"
+license = "Apache-2.0"
+repository = "https://github.com/awslabs/smithy-rs"
+
+
+[features]
+hickory-dns = ["dep:hickory-resolver", "dep:tokio", "tokio/rt"]
+
+
+[dependencies]
+aws-smithy-runtime-api = { path = "../aws-smithy-runtime-api", features = ["client"]}
+tokio = { version = "1.40.0", features = [], optional = true}
+
+
+# hickory-resolver enables some tokio features that break in wasm
+[target.'cfg(not(target_family = "wasm"))'.dependencies]
+hickory-resolver = {version = "0.25.2", optional = true}
+
+
+[dev-dependencies]
+criterion = "0.5.1"
+
+
+
+[package.metadata.docs.rs]
+all-features = true
+targets = ["x86_64-unknown-linux-gnu"]
+cargo-args = ["-Zunstable-options", "-Zrustdoc-scrape-examples"]
+rustdoc-args = ["--cfg", "docsrs"]
+# End of docs.rs metadata
