+++ rust-runtime/aws-smithy-dns/README.md
@@ -1,0 +1,7 @@
+# aws-smithy-dns
+
+Built-in DNS resolver implementations for smithy-rs clients.
+
+<!-- anchor_start:footer -->
+This crate is part of the [AWS SDK for Rust](https://awslabs.github.io/aws-sdk-rust/) and the [smithy-rs](https://github.com/smithy-lang/smithy-rs) code generator. In most cases, it should not be used directly.
+<!-- anchor_end:footer -->
