+++ rust-runtime/aws-smithy-runtime/Cargo.toml
@@ -18,7 +18,6 @@ connector-hyper-0-14-x = ["dep:aws-smithy-http-client", "aws-smithy-http-client?
 tls-rustls = ["dep:aws-smithy-http-client", "aws-smithy-http-client?/legacy-rustls-ring", "connector-hyper-0-14-x"]
 default-https-client = ["dep:aws-smithy-http-client", "aws-smithy-http-client?/rustls-aws-lc"]
 rt-tokio = ["tokio/rt"]
-hickory-dns = ["dep:hickory-resolver", "rt-tokio"]
 
 # Features for testing
 # NOTE: `test-util` originally enabled the HTTP related test utils which have since been relocated to `aws-smithy-http-client`
@@ -66,10 +65,6 @@ pin-utils = "0.1.0"
 tokio = { version = "1.40.0", features = [] }
 tracing = "0.1.40"
 tracing-subscriber = { version = "0.3.16", optional = true, features = ["env-filter", "fmt", "json"] }
-
-# hickory-resolver enables some tokio features that break in wasm
-[target.'cfg(not(target_family = "wasm"))'.dependencies]
-hickory-resolver = {version = "0.25.2", optional = true}
 
 [dev-dependencies]
 approx = "0.5.1"
